requirements.txt
cryptography>=42.0.0
jsonschema>=4.21.0
pyyaml>=6.0.1
requests>=2.31.0
pytest>=8.0.0
pytest-cov>=4.1.0
